Allow for URL exceptions to be given
Previous to this commit, there was no way to specify specific URLs or
URL patterns that will be except from requiring authentication.  This
commit introduces a configuration parameter,
:allow_unauthenticated_urls, that can be an array of regular expressions
or strings that will be used to match the requesting URL.  If the
requesting URL matches any of the exceptions given, the request will be
allowed through.