Use REQUEST_URI instead of PATH_INFO
Previous to this commit, the PATH_INFO header information was used to
determine the calling page.  This however doesn't always give the
exptected result.  For example, for rack applications launched by
passenger using RackBaseURI, the path to the rack application is
stripped from PATH_INFO.

This commit uses the REQUEST_URI to resolve the issue